backend/requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.75.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
iso8601==0.1.16
isort==5.10.1
mypy-extensions==0.4.3
orjson==3.6.7
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
pypika-tortoise==0.1.3
python-jose==3.3.0
pytz==2022.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
tortoise-orm==0.18.1
typing-extensions==4.1.1
uvicorn==0.17.6
win32-setctime==1.1.0
